added dht11 sensor and changed dust sensor hookup
this version of the firmware works differently then before. i somehow screwed my serial line communication from the feather to the dust sensor and i also saw, that even when sleeping the dust sensor was consuming 6ma - way too much. so i added a new wiring diagram, used a pololu buck/boost converter with an enable line that we can use, to completely turn the dust sensor off. this now allows to not use serial but to use a simple voltage divider to read from pwm siglas the dust level. we loose 1 decimal of precision but i don't think we are that accurat anyway. the circuit now only needs .5 mA when sleeping which allows the device to sleep for 166 days. as the consumption of the sensor running 30 seconds needs much more power (can only run 2 weeks when running for 30 seconds every 5 minutes) this is good enough. i modified the upload so data received now includes the temperature and the humidity: December 25, 2017 at 05:46PM	alldata	[M 4.34375]	[2 17 19 47]